Make SDL2 relative mouse sensitivity match `MouseHandler` sensitivity on Windows
This is needed on Windows so that the SDL cursor matches framework's cursor for smoother
transitions when toggling relative mode.

If this were implemented for all platforms, we'd have scaling applied two times.
On Windows, we have our own raw input handler so we scaling SDL relative mouse events
doesn't affect us.